refactor: unify theme provider
