--- client/src/components/ThemeProvider.jsx
@@ -1,22 +1,0 @@
-// client/src/components/ThemeProvider.jsx
-import { useSelector } from 'react-redux';
-import { useEffect } from 'react';
-
-export default function ThemeProvider({ children }) {
-    const { theme } = useSelector((state) => state.theme);
-
-    useEffect(() => {
-        const root = window.document.documentElement;
-        root.classList.remove('light', 'dark');
-        root.classList.add(theme);
-        localStorage.setItem('theme', theme);
-    }, [theme]);
-
-    return (
-        <div className={theme}>
-            <div className='bg-white text-gray-700 dark:text-gray-200 dark:bg-[rgb(16,23,42)] min-h-screen'>
-                {children}
-            </div>
-        </div>
-    );
-}
