--- client/src/ui/ThemeProvider.jsx
@@ -1,26 +1,0 @@
-/* eslint-disable react/prop-types */
-import { createContext, useContext, useEffect, useState } from 'react';
-
-const ThemeContext = createContext({ theme: 'light', setTheme: () => {} });
-
-export function ThemeProvider({ children }) {
-  const prefersDark = typeof window !== 'undefined' && window.matchMedia('(prefers-color-scheme: dark)').matches;
-  const [theme, setTheme] = useState(() => localStorage.getItem('theme') || (prefersDark ? 'dark' : 'light'));
-
-  useEffect(() => {
-    const root = document.documentElement;
-    root.classList.remove('light', 'dark');
-    root.classList.add(theme);
-    localStorage.setItem('theme', theme);
-  }, [theme]);
-
-  return (
-    <ThemeContext.Provider value={{ theme, setTheme }}>
-      {children}
-    </ThemeContext.Provider>
-  );
-}
-
-export function useTheme() {
-  return useContext(ThemeContext);
-}
