+++ client/src/ui/ThemeToggle.jsx
@@ -1,11 +1,11 @@
-import { useTheme } from './ThemeProvider';
+import { useTheme } from '../theme/ThemeProvider.jsx';
 
 export default function ThemeToggle() {
-  const { theme, setTheme } = useTheme();
+  const { theme, toggle } = useTheme();
   return (
     <button
       type="button"
-      onClick={() => setTheme(theme === 'dark' ? 'light' : 'dark')}
+      onClick={toggle}
       className="p-2 rounded-md border border-border bg-surface text-text hocus:bg-muted/10"
     >
       {theme === 'dark' ? '🌙' : '☀️'}
